+++ src/containers/team/Modal/type.ts
@@ -1,5 +1,11 @@
 import { Team } from '@/types';
 
+export interface TeamModalProps {
+  teamInfo?: Team;
+  isOpen: boolean;
+  onClose: () => void;
+}
+
 export interface TeamActionModalProps extends Pick<Team, 'id' | 'name'> {
   isOpen: boolean;
   onClose: () => void;
