--- src/containers/team/TeamModal/type.ts
@@ -1,7 +1,0 @@
-import { Team } from '@/types';
-
-export interface TeamModalProps {
-  teamInfo?: Team;
-  isOpen: boolean;
-  onClose: () => void;
-}
